+++ include/miopen/miopen.h
@@ -8236,10 +8236,10 @@ MIOPEN_EXPORT miopenStatus_t miopenRoIAlignBackward(miopenHandle_t handle,
                                                     const void* rois,
                                                     const miopenTensorDescriptor_t inputGradDesc,
                                                     void* inputGrad,
-                                                    const int32_t alignedHeight,
-                                                    const int32_t alignedWidth,
+                                                    const uint64_t alignedHeight,
+                                                    const uint64_t alignedWidth,
                                                     const float spatialScale,
-                                                    const int32_t samplingRatio,
+                                                    const int64_t samplingRatio,
                                                     const bool aligned);
 
 /** @}*/
